LU-7934 osp: fix tr->otr_next_id overflow
The tracker next_id and current_id u32 type was based on
max llog records. But llog use cyclic store for records, so
llog could store infinite number of records and limited by
max number at moument of time. The u32 type could be
overflowed easy if a server isn`t rebooted.
osp_sync_id_get()) snx11126-OST0045-osc-MDT0000: next 0,
	last synced 4294967205
This fix changes type u32 to u64 for *id. Now, we store only
low part current_id to llog record header id. And restore the
full u64 from record header later. It is possible because
llog catalog can store 64768^2 and it is less than u32 max value.

The patch adds test to check u32 overflow for otr_next_id field
of osp_id_tracker.

Recreate the next assertion
LustreError: 185667:0:(osp_sync.c:1544:osp_sync_id_get())
	snx11126-OST0045-osc-MDT0000: next 0, last synced 4294967205
LustreError: 231396:0:(osp_sync.c:1545:osp_sync_id_get()) LBUG

Seagate-bug-id: MRP-3367
Reviewed-on: http://review.whamcloud.com/19190
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ lustre/osp/osp_internal.h
@@ -55,8 +55,8 @@
  */
 struct osp_id_tracker {
 	spinlock_t		 otr_lock;
-	__u32			 otr_next_id;
-	__u32			 otr_committed_id;
+	__u64			 otr_next_id;
+	__u64			 otr_committed_id;
 	/* callback is register once per diskfs -- that's the whole point */
 	struct dt_txn_callback	 otr_tx_cb;
 	/* single node can run many clusters */
@@ -215,12 +215,12 @@ struct osp_device {
 	/* osd api's commit cb control structure */
 	struct dt_txn_callback		 opd_syn_txn_cb;
 	/* last used change number -- semantically similar to transno */
-	unsigned long			 opd_syn_last_used_id;
+	__u64				 opd_syn_last_used_id;
 	/* last committed change number -- semantically similar to
 	 * last_committed */
-	unsigned long			 opd_syn_last_committed_id;
+	__u64				 opd_syn_last_committed_id;
 	/* last processed (taken from llog) id */
-	unsigned long			 opd_syn_last_processed_id;
+	__u64				 opd_syn_last_processed_id;
 	struct osp_id_tracker		*opd_syn_tracker;
 	struct list_head		 opd_syn_ontrack;
 	/* stop processing new requests until barrier=0 */
@@ -420,7 +420,7 @@ static inline struct osp_thread_info *osp_env_info(const struct lu_env *env)
 }
 
 struct osp_txn_info {
-	__u32   oti_current_id;
+	__u64   oti_current_id;
 };
 
 extern struct lu_context_key osp_txn_key;
